Wrap menu texts, check samples, show killing sprees
Create a text wrapping functionality to the menu text component. Wrap
the client error texts and texts in the connection status menu.

Check the received sample number in the client before trying to play it.
That allows using new samples in future versions without breaking the
old ones.

Play the killing spree sound only in the attacker’s client. Show a
message when a player is on a killing spree and when the killing spree
ends.