Record initial data, revert position recording
Record the server settings, welcome message and more player data at the
start of a replay. Record settings when reloaded.

Put back the recording of players’ positions in all frames because the
extrapolation does not correctly work when replaying.

Handle pausing better. Stop the replay if no more data could be read.
Play rocket sounds in replays. Do not play room specific sounds if the
room is not visible.

Stop generating deathbringer smoke when the replay is paused. Draw
player names regardless of their alpha in replays.