Fix copying of BinaryBuffer
Fix copying of BinaryBuffer. The implicit copying of BinaryWriter led
to both writers pointing to memory in the source buffer. Fixes at least
the sending of data_gameover_show and data_movements_shots in extended
format when it differs from the old one, i.e. with large numbers to
send.