Have bots capture on near-dropped flags
When calculating whether a dropped flag is at base to be captured on,
accept all flags that are near enough to be captured on without help
according to the new rules. Allows bots to make more captures when
carrying own team flags is enabled (so that a carrier can’t return the
flag to be captured on) and there is no teammate to carry the other
flag.