Improve bot in-room navigation
When the targeted location is behind a wall, see if it is directly
reachable from some of the 8 points reached by traveling in a basic
direction as far as possible.

Fixes bots getting temporarily stuck in many types of rooms with simple
obstructions. More maze-like rooms are still a problem.