Have bots avoid enemy bases
Penalize going to the enemy base in the same way as respawn areas. When
there are alternatives almost as near, makes the bots go to them instead
of the enemy base. They show more interest toward wild flags and
defense, and avoid going through enemy bases when there are alternative
routes.