Bots: Treat dropped flags in base as unmoved
When choosing where to go on the map, have bots treat flags that are
dropped within their base area the same as if they were at the base
unmoved. When in the same room with the flag, keep the old behavior.

Eliminates some superfluous running back to return flags that are safe
enough.