Create protocol extension for server settings
Create a protocol extension to send more complete data for server
settings. Allows new clients to read correct settings and show them
on the server settings menu. Fixes also a crash when the server sends
settings with values that do not fit in a single byte.

Increase the replay version as the server settings are stored in
replays in the new format.