Use protocol extension for gameover data
Use the protocol extension level 4 for the data_gameover_show message
that contains the capture limit and time limit. Also limit the values
sent to old clients to be in a valid range (single byte in this case).

Fixes a crash on the server side when a game ends and the capture limit,
time limit or extra time is more than 255 captures or minutes.