Fix bot crash with random_wild_flag
Fix a crash in bot code that occurs when there is a flag without a
corresponding base, caused by the setting random_wild_flag.

Changes the regular behavior as a side-effect: Bots now consider a
flag in a different base of its color to be at base too, also when
considering which flags to ignore at the moment.