Fix protocol extensions in replay playback
Fix the use of protocol extensions in replay playback. Set the protocol
extensions level in the client according to the replay file version.
Previously it was never set.

Fixes "Format error in replay file" and other errors when the replay
file version clashes with the version on the last server played on, or
at random if replaying without having connected to a server before.