Fix keyboard aiming with smooth turning option
When the keyboard with smooth turning is selected for the preferred aim
option and the server settings do not allow free turning, the player can
not move the gun to left and right. Fix it by checking that the free
turning setting is on before clearing left and right from the player’s
controls.