Fix bots stuck where flag was seen
Fix bots getting stuck for up to 5 seconds in a room where an enemy
carrier was predicted to enter when they got hidden. In that situation,
go where the carrier actually was when last seen. Existing code will
mark them unseen and go elsewhere if not there either.

Fix a minor one-frame delay in some actions when a hidden enemy is lost
track of.