Fix replay menu selections with adds/removes
Rewrite TextTree/TreeItem item selection code in a way that works with
addition and removal of items after moving the initial selection.

Fixes the effect of multiple lines being selected at once and/or the
selected replay changing, and even possible crashes, in the replay menu
when reopened after replays are added or deleted.