Cache old partial replays too
Store information about old partial replays in the cache as well as
finished ones. Since the cache is used to determine deletions for the
replay menu, this fixes deleted partial replays not being removed from
the menu.

Since new partial replays don't appear to the client anymore due to the
suffix .ongoing, this doesn't lead to using outdated cached information,
which was the original motivation for skipping partial replays.