Bots: Defend carriers in neighboring room
Have bots defend carriers on the other side of a room border when
already located there, if there are no enemies or other flags in the
area.

Does not significantly affect the performance of a pure bot team
(at least not against other bots) but makes them nicer to co-operate
with.