Tests: Add disabled check for mapgen return value
Add to the mapgen test program a check that MapGenerator returns some
"good" results with each sensible set of parameters. Unfortunately, by
default these checks have to be disabled due to the required
multiplication in runtime.

Remove a spurious bitmask++ from the repeat loop, which was causing one
bitmask to be skipped, and made the set number of repeats insignificant.