Mapgen: Make asymmetric maps green base first
In MapGenerator, choose bases for an asymmetric map with a green flag in
a new way:
1. Pick a random room for the green flag.
2. Select a pair of team bases with an equal distance from the green
   flag that maximizes a combination of the distances between pairs of
   bases.

Replaces the old method of choosing team bases with a maximum distance
between them first, and then trying to find an equidistant green base.
This generated maps that were unnecessarily similar to each other, and
often failed to find a valid position for the green flag forcing a
restarted generation.